+++ app/src/main/java/com/juniori/puzzle/di/AuthModule.kt
@@ -1,7 +1,7 @@
 package com.juniori.puzzle.di
 
 import com.google.firebase.auth.FirebaseAuth
-import com.juniori.puzzle.data.auth.AuthRepository
+import com.juniori.puzzle.domain.repository.AuthRepository
 import com.juniori.puzzle.data.auth.AuthRepositoryImpl
 import dagger.Module
 import dagger.Provides
