+++ app/src/main/java/com/juniori/puzzle/domain/usecase/GetSearchedMyVideoUseCase.kt
@@ -1,6 +1,6 @@
 package com.juniori.puzzle.domain.usecase
 
-import com.juniori.puzzle.data.video.VideoRepository
+import com.juniori.puzzle.domain.repository.VideoRepository
 import com.juniori.puzzle.di.MockData
 import javax.inject.Inject
