+++ app/src/main/java/com/juniori/puzzle/domain/usecase/GetUserInfoUseCase.kt
@@ -1,8 +1,6 @@
 package com.juniori.puzzle.domain.usecase
 
-import com.juniori.puzzle.data.Resource
-import com.juniori.puzzle.data.auth.AuthRepository
-import com.juniori.puzzle.domain.entity.UserInfoEntity
+import com.juniori.puzzle.domain.repository.AuthRepository
 import javax.inject.Inject
 
 class GetUserInfoUseCase @Inject constructor(private val authRepository: AuthRepository) {
