+++ app/src/main/java/com/juniori/puzzle/domain/usecase/RequestLoginUseCase.kt
@@ -1,7 +1,7 @@
 package com.juniori.puzzle.domain.usecase
 
 import com.google.android.gms.auth.api.signin.GoogleSignInAccount
-import com.juniori.puzzle.data.auth.AuthRepository
+import com.juniori.puzzle.domain.repository.AuthRepository
 import javax.inject.Inject
 
 class RequestLoginUseCase @Inject constructor(private val authRepository: AuthRepository) {
