+++ app/src/main/java/com/juniori/puzzle/domain/usecase/RequestLogoutUseCase.kt
@@ -1,6 +1,6 @@
 package com.juniori.puzzle.domain.usecase
 
-import com.juniori.puzzle.data.auth.AuthRepository
+import com.juniori.puzzle.domain.repository.AuthRepository
 import javax.inject.Inject
 
 class RequestLogoutUseCase @Inject constructor(private val authRepository: AuthRepository) {
